GPII-4022: Committed broken work in progress on "resource loading" refactor to get input.
