+++ index.js
@@ -8,7 +8,6 @@ fluid.module.register("gpii-json-schema", __dirname, require);
 // Require all of the server-side components at once.
 require("./src/js/common/gss-metaschema");
 require("./src/js/common/validator");
-require("./src/js/common/validation-errors");
 require("./src/js/common/schemaHolder");
 require("./src/js/common/schemaValidatedComponent");
 require("./src/js/common/schemaValidatedModelComponent");
