+++ package.json
@@ -17,27 +17,27 @@
     "author": "Tony Atkins <tony@raisingthefloor.org>",
     "license": "BSD-3-Clause",
     "dependencies": {
-        "ajv": "6.10.0",
+        "ajv": "6.10.2",
         "gpii-binder": "1.0.5",
         "gpii-express": "1.0.15",
-        "gpii-handlebars": "1.1.4",
-        "infusion": "3.0.0-dev.20190507T155813Z.4781871fd.FLUID-6148",
+        "gpii-handlebars": "2.1.0-dev.20190923T144705Z.38550ff.GPII-4100",
+        "infusion": "3.0.0-dev.20190926T213421Z.e84b3cf0e.FLUID-6148",
         "kettle": "1.11.0"
     },
     "devDependencies": {
-        "eslint": "6.0.0",
+        "eslint": "6.5.0",
         "eslint-config-fluid": "1.3.0",
         "foundation-sites": "6.4.1",
         "gpii-grunt-lint-all": "1.0.5",
-        "gpii-testem": "2.1.10-dev.20190404T122608Z.b51705e.GPII-3457",
+        "gpii-testem": "2.1.10",
         "grunt": "1.0.4",
-        "handlebars": "4.1.2",
-        "markdown-it": "8.4.2",
+        "handlebars": "4.4.0",
+        "markdown-it": "10.0.0",
         "mkdirp": "0.5.1",
         "node-jqunit": "1.1.8",
-        "nyc": "12.0.2",
+        "nyc": "14.1.1",
         "request": "2.88.0",
-        "rimraf": "2.6.3",
+        "rimraf": "3.0.0",
         "testem": "2.17.0"
     }
 }
