+++ tests/browser-fixtures/errorBinder-tests.html
@@ -2,23 +2,10 @@
     <head>
         <title>"Error Binder" Tests</title>
 
-        <!-- jQuery -->
-        <script type="text/javascript" src="../../node_modules/infusion/src/lib/jquery/core/js/jquery.js"></script>
+        <!-- Bring in client library being tested -->
 
         <!-- Fluid dependencies -->
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidDocument.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/Fluid.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidPromises.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidDOMUtilities.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidIoC.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/DataBinding.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidView.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/lib/fastXmlPull/js/fastXmlPull.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/renderer/js/fluidParser.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/renderer/js/fluidRenderer.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/FluidRequests.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/ModelTransformation.js"></script>
-        <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/ModelTransformationTransforms.js"></script>
+        <script type="text/javascript" src="../../node_modules/infusion/dist/infusion-all.js"></script>
         <script type="text/javascript" src="../../node_modules/infusion/src/framework/core/js/MessageResolver.js"></script>
 
         <!-- IoC Test Framework dependencies -->
@@ -45,12 +32,14 @@
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/md-client.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/common/renderer.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/renderer.js"></script>
+        <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/serverResourceAware.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/templateAware.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/templateFormControl.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/templateManager.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/templateMessage.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/templateRequestAndRender.js"></script>
         <script type="text/javascript" src="../../node_modules/gpii-handlebars/src/js/client/hasRequiredOptions.js"></script>
+
 
         <!-- gpii-binder dependencies -->
         <script type="text/javascript" src="../../node_modules/gpii-binder/src/js/binder.js"></script>
@@ -74,8 +63,6 @@
         <script src="/src/js/common/gss-metaschema.js"></script>
         <script src="/src/js/common/schemaValidatedComponent.js"></script>
         <script src="/src/js/common/schemaValidatedModelComponent.js"></script>
-        <script src="/src/js/common/validation-errors.js"></script>
-        <script src="/src/js/common/hashString.js"></script>
         <script src="/src/js/common/orderedStringify.js"></script>
         <script src="/src/js/common/validator.js"></script>
