+++ tests/browser-fixtures/schema-validated-component-tests.html
@@ -17,7 +17,6 @@
 
         <!-- code under test -->
         <script src="/src/js/common/gss-metaschema.js"></script>
-        <script src="/src/js/common/validation-errors.js"></script>
         <script src="/src/js/common/validator.js"></script>
         <script src="/src/js/common/schemaValidatedComponent.js"></script>
     </head>
