+++ tests/browser-fixtures/validator-global-component-tests.html
@@ -19,7 +19,6 @@
 
         <!-- code under test -->
         <script src="/src/js/common/gss-metaschema.js"></script>
-        <script src="/src/js/common/validation-errors.js"></script>
         <script src="/src/js/common/orderedStringify.js"></script>
         <script src="/src/js/common/validator.js"></script>
     </head>
