+++ tests/browser-fixtures/validator-static-function-tests.html
@@ -17,7 +17,6 @@
 
         <!-- code under test -->
         <script src="/src/js/common/gss-metaschema.js"></script>
-        <script src="/src/js/common/validation-errors.js"></script>
         <script src="/src/js/common/validator.js"></script>
     </head>
     <body>
@@ -29,6 +28,8 @@
 
         <!-- The tests -->
         <script src="../js/common/validator-tests-ajv-errors-testDefs.js"></script>
-        <script src="../js/common/validator-static-function-tests.js"></script>
+        <script src="../js/common/validator-static-function-testDefs.js"></script>
+
+        <script src="js/messageLoadingRunner.js"></script>
     </body>
 </html>
